Move input actions handling to the separate file
